third_party/WebKit/LayoutTests/fragmentation/become-fragmented-same-widths.html

<!DOCTYPE HTML>
<!-- A block with fixed width is 220px tall. When turning its container into a
     multicol container with a column height of 300px, we need to re-lay out the
     block, even if it would seem to fit in one column (220px < 300px). It
     contains a forced break. -->
<p>There should be a blue square below.</p>
<div id="multicol" style="position:relative; column-gap:0; column-fill:auto; width:100px; height:300px;">
    <div style="width:50px;">
        <div id="block1" style="height:100px; background:blue;"></div>
        <div style="height:20px;"></div>
        <div id="block2" style="break-before:column; height:100px; background:blue;"></div>
    </div>
</div>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    test(() => {
        document.body.offsetTop;
        document.getElementById("multicol").style.columns = "2";
        var block1 = document.getElementById("block1");
        var block2 = document.getElementById("block2");
        assert_equals(block1.offsetTop, 0);
        assert_equals(block1.offsetLeft, 0);
        assert_equals(block2.offsetTop, 0);
        assert_equals(block2.offsetLeft, 50);
    }, "Need relayout when becoming fragmented, even if width is the same.");
</script>
